requirements.txt
ruaccent
pyarrow
qwen_omni_utils
jiwer
pyyaml
